добавила новое поле в рецептах,удалила поле номер телефона распределила контроллеры и изменила метод регистрации
