+++ src/main/java/com/finalka/controller/UserController.java
@@ -7,6 +7,7 @@ import com.finalka.enums.Units;
 import com.finalka.service.MenuService;
 import com.finalka.service.RecipesService;
 import com.finalka.service.ReviewService;
+import com.finalka.service.impl.ReminderServiceImpl;
 import com.finalka.service.impl.UserServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
@@ -14,6 +15,8 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.apache.velocity.exception.ResourceNotFoundException;
@@ -29,7 +32,7 @@ import java.util.Map;
 
 import static org.hibernate.query.sqm.tree.SqmNode.log;
 
-
+@Tag(name = "MagicMenu", description = "Тут находятся все роуты для пользователей")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/users")
@@ -38,6 +41,7 @@ public class UserController {
     private final RecipesService recipeService;
     private final MenuService menuService;
     private final ReviewService reviewService;
+    private final ReminderServiceImpl reminderService;
 
 
     @ApiResponses(value = {
@@ -82,6 +86,7 @@ public class UserController {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
+
     @Operation(summary = "Этот роут для поиска рецепта по нескольким продуктам")
     @ApiResponses(value = {
             @ApiResponse(
@@ -99,6 +104,7 @@ public class UserController {
         List<RecipeWithProductDTO> recipes = recipeService.findRecipesByProducts(userProducts);
         return ResponseEntity.ok(recipes);
     }
+
     @Operation(summary = "Этот роут возвращает количество всех продуктов в одном меню по айди")
     @ApiResponses(value = {
             @ApiResponse(
@@ -162,8 +168,7 @@ public class UserController {
             return new ResponseEntity<>("Internal server error", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
-
-    @Operation(summary = "Этот роут возвращает отзыв к рецепту по его айди")
+    @Operation(summary = "Этот роут для установки напоминание")
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
@@ -174,17 +179,28 @@ public class UserController {
                     responseCode = "404",
                     description = "Не найдено")
     })
-
-    @GetMapping("/recipe/{recipeId}")
-    public ResponseEntity<List<ReviewDTO>> getReviewsByRecipeId(@PathVariable Long recipeId) {
-        try {
-            List<ReviewDTO> reviews = reviewService.getReviewsByRecipeId(recipeId);
-            return new ResponseEntity<>(reviews, HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+    @PostMapping("/setReminder")
+    public ResponseEntity<String> createReminder(@RequestParam int hour, @RequestParam int minute, @RequestParam String message) {
+        reminderService.setReminder(hour, minute, message);
+        return ResponseEntity.ok("Напоминание успешно установлено");
     }
-    @Operation(summary = "Этот роут для добовление рецептов в меню")
+    @Operation(summary = "Этот роут возвращает все напоминание")
+    @ApiResponses(value = {
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "Успешная операция",
+                    content = {@Content(mediaType = "application/json",
+                            array = @ArraySchema(schema = @Schema(implementation = ReminderDto.class)))}),
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "Не найдено")
+    })
+    @GetMapping("/getAllReminders")
+    public ResponseEntity<List<ReminderDto>> getAllReminders() {
+        List<ReminderDto> reminders = reminderService.getAllReminders();
+        return ResponseEntity.ok(reminders);
+    }
+    @Operation(summary = "Этот роут отменяет  напоминание по айди пользователя ")
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
@@ -195,36 +211,84 @@ public class UserController {
                     responseCode = "404",
                     description = "Не найдено")
     })
-    @PostMapping("/add-to-menu")
-    public ResponseEntity<String> addRecipeToMenu(@RequestBody RecipeAddProductDto menuRecipeRequestDto) {
-        try {
-            recipeService.addRecipeToMenu(menuRecipeRequestDto.getMenuId(), menuRecipeRequestDto.getRecipeId());
-            return ResponseEntity.ok("Рецепт успешно добавлен в меню");
-        } catch (Exception e) {
-            log.error("Ошибка при добавлении рецепта в меню", e);
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Не удалось добавить рецепт в меню");
-        }
+    @DeleteMapping("/cancel/{reminderId}")
+    public ResponseEntity<String> cancelReminder(@PathVariable Long reminderId) {
+        reminderService.cancelReminder(reminderId);
+        return ResponseEntity.ok("Напоминание успешно отменено");
     }
 
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "В базе есть доступные рецепты",
+                    description = "Напоминание найден и успешно обновлен",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = MenuDTO.class))}),
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "Напоминание не найден")
+    })
+    @Operation(summary = "Роут обновляет напоминание, id напоминание передается непосредственно в модели," +
+            " по ней и идет поиск, важно, что бы все поля не были пустыми иначе засетит null, но передавать создателя и " +
+            "обновляющего с датами не нужно, это делает бэк")
+    @PutMapping("/update/{reminderId}")
+    public ResponseEntity<Void> updateReminder(@PathVariable Long reminderId, @RequestParam int hour, @RequestParam int minute, @RequestParam String message) {
+        CreateReminderDto reminderDto = new CreateReminderDto();
+        reminderDto.setHour(hour);
+        reminderDto.setMinute(minute);
+        reminderDto.setMessage(message);
+
+        try {
+            reminderService.updateReminder(reminderId, reminderDto);
+            return ResponseEntity.ok().build();
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+    }
+
+    @Operation(summary = "Роут удаляет напоминание по id")
+    @ApiResponses(value = {
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "Напоминание найден и успешно удален",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = String.class))}),
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "Напоминание не найден")
+    })
+    @DeleteMapping("/delete/{reminderId}")
+    public ResponseEntity<Void> deleteReminder(@PathVariable Long reminderId) {
+        reminderService.deleteReminder(reminderId);
+        return ResponseEntity.ok().build();
+    }
+
+    @ApiResponses(value = {
+            @ApiResponse(
+                    responseCode = "200",
+                    description = "Получены все напоминание пользователя",
                     content = {@Content(mediaType = "application/json",
                             array = @ArraySchema(schema = @Schema(implementation = RecipesDto.class)))}),
             @ApiResponse(
                     responseCode = "404",
-                    description = "Рецепта нет")
+                    description = "Напоминаний нет")
     })
-    @Operation(summary = "Роут возвращает все не удаленные рецепты")
-    @GetMapping("/all")
-    public ResponseEntity<List<RecipesDto>> findAll() {
+    @Operation(summary = "Роут возвращает все свои напоминание пользователя")
+    @GetMapping("/getMyReminders")
+    public ResponseEntity<List<ReminderDto>> getUserReminders(@RequestHeader("Authorization") String token) {
         try {
-            return new ResponseEntity<>(recipeService.findAll(), HttpStatus.OK);
-        } catch (NullPointerException nullPointerException) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            if (token == null || !token.startsWith("Bearer ")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            }
+
+            String authToken = token.substring(7);
+
+            List<ReminderDto> reminders = reminderService.getUserReminders();
+
+            return ResponseEntity.ok(reminders);
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            log.error("Ошибка при получении напоминаний пользователя", e);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
     }
 }
+
