+++ src/main/java/com/finalka/security/SecurityConfig.java
@@ -55,9 +55,11 @@ public class SecurityConfig {
                         "/swagger-ui/**", "/v3/api-docs/**", "/swagger-ui.html",
                         "/swagger-resources/", "/webjars/")
                 .permitAll();
-        http.authorizeHttpRequests().requestMatchers(HttpMethod.POST, "/api/registration").permitAll();
+        http.authorizeHttpRequests().requestMatchers(HttpMethod.POST, "/api/all").permitAll();
         http.authorizeHttpRequests().requestMatchers("/api/chefs/**").hasAuthority( "CHEF");
         http.authorizeHttpRequests().requestMatchers("/api/users/**").hasAuthority( "USER");
+        http.authorizeHttpRequests().requestMatchers("/api/admins/**").hasAuthority( "ADMIN");
+        http.authorizeHttpRequests().requestMatchers("/api/comminis/**").hasAnyAuthority( "ADMIN","CHEF","USER");
         http.authorizeHttpRequests().anyRequest().authenticated();
         http.apply(CustomSecurityDetails.customDsl(userRepo));
         http.addFilterBefore(new CustomAuthorizationFilter(), UsernamePasswordAuthenticationFilter.class);
