+++ src/main/java/com/finalka/service/impl/RecipeServiceImpl.java
@@ -93,6 +93,7 @@ public class RecipeServiceImpl implements RecipesService {
                 .id(recipe.getId())
                 .nameOfFood(recipe.getNameOfFood())
                 .description(recipe.getDescription())
+                .portion(recipe.getPortion())
                 .imageBase64(recipe.getImageBase64())
                 .linkOfVideo(recipe.getLinkOfVideo())
                 .quantityOfProduct(recipe.getQuantityOfProduct())
@@ -125,6 +126,7 @@ public class RecipeServiceImpl implements RecipesService {
             Recipes recipe = Recipes.builder()
                     .nameOfFood(recipeDTO.getNameOfFood())
                     .description(recipeDTO.getDescription())
+                    .portion(recipeDTO.getPortion())
                     .imageBase64(recipeDTO.getImageBase64())
                     .linkOfVideo(recipeDTO.getLinkOfVideo())
                     .quantityOfProduct(recipeDTO.getQuantityOfProduct())
