+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 # Datasource configuration
-spring.datasource.url=jdbc:postgresql://localhost:5432/finalka
+spring.datasource.url=jdbc:postgresql://localhost:5432/final
 spring.datasource.username=postgres
 spring.datasource.password=postgres
 spring.datasource.driver-class-name=org.postgresql.Driver
